squirrel and scany

+++ internal/inputs/search_params.go
@@ -3,8 +3,8 @@ package inputs
 import "time"
 
 type SearchParams struct {
-	Offset int
-	Limit  int
+	Offset uint64
+	Limit  uint64
 
 	CreatedAtFrom *time.Time
 	CreatedAtTo   *time.Time
